+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+- Add partition_compaction_key in root of event for log compacted event type
+
 ## [2.8.2] - 2018-07-31
 
 ### Removed