+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Removed partition key fields check
 
+### Fixed
+- Fixed and added check for schema root object type to be only of type object.
+
 ## [2.8.2] - 2018-07-31
 
 ### Removed